+++ actix-tls/src/openssl.rs
@@ -6,10 +6,11 @@ use std::task::{Context, Poll};
 use actix_codec::{AsyncRead, AsyncWrite};
 use actix_service::{Service, ServiceFactory};
 use actix_utils::counter::{Counter, CounterGuard};
-use futures_util::future::{ok, FutureExt, LocalBoxFuture, Ready};
+use futures_util::future::{ready, Ready};
+use futures_util::ready;
 
-pub use open_ssl::ssl::{AlpnError, SslAcceptor, SslAcceptorBuilder};
-pub use tokio_openssl::{HandshakeError, SslStream};
+pub use open_ssl::ssl::{AlpnError, Error, Ssl, SslAcceptor};
+pub use tokio_openssl::SslStream;
 
 use crate::MAX_CONN_COUNTER;
 
@@ -45,7 +46,7 @@ impl<T: AsyncRead + AsyncWrite> Clone for Acceptor<T> {
 impl<T: AsyncRead + AsyncWrite + Unpin + 'static> ServiceFactory for Acceptor<T> {
     type Request = T;
     type Response = SslStream<T>;
-    type Error = HandshakeError<T>;
+    type Error = Error;
     type Config = ();
     type Service = AcceptorService<T>;
     type InitError = ();
@@ -53,11 +54,11 @@ impl<T: AsyncRead + AsyncWrite + Unpin + 'static> ServiceFactory for Acceptor<T>
 
     fn new_service(&self, _: ()) -> Self::Future {
         MAX_CONN_COUNTER.with(|conns| {
-            ok(AcceptorService {
+            ready(Ok(AcceptorService {
                 acceptor: self.acceptor.clone(),
                 conns: conns.clone(),
                 io: PhantomData,
-            })
+            }))
         })
     }
 }
@@ -71,7 +72,7 @@ pub struct AcceptorService<T> {
 impl<T: AsyncRead + AsyncWrite + Unpin + 'static> Service for AcceptorService<T> {
     type Request = T;
     type Response = SslStream<T>;
-    type Error = HandshakeError<T>;
+    type Error = Error;
     type Future = AcceptorServiceResponse<T>;
 
     fn poll_ready(&mut self, ctx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
@@ -83,31 +84,52 @@ impl<T: AsyncRead + AsyncWrite + Unpin + 'static> Service for AcceptorService<T>
     }
 
     fn call(&mut self, req: Self::Request) -> Self::Future {
-        let acc = self.acceptor.clone();
-        AcceptorServiceResponse {
-            _guard: self.conns.get(),
-            fut: async move {
-                let acc = acc;
-                tokio_openssl::accept(&acc, req).await
-            }
-            .boxed_local(),
-        }
+        let guard = self.conns.get();
+        let stream = self.ssl_stream(req);
+        AcceptorServiceResponse::Init(Some(stream), Some(guard))
     }
 }
 
-pub struct AcceptorServiceResponse<T>
+impl<T: AsyncRead + AsyncWrite + Unpin + 'static> AcceptorService<T> {
+    // construct a new SslStream.
+    // At this point the SslStream does not perform any IO.
+    // The handshake would happen later in AcceptorServiceResponse
+    fn ssl_stream(&self, stream: T) -> Result<SslStream<T>, Error> {
+        let ssl = Ssl::new(self.acceptor.context())?;
+        let stream = SslStream::new(ssl, stream)?;
+        Ok(stream)
+    }
+}
+
+pub enum AcceptorServiceResponse<T>
 where
     T: AsyncRead + AsyncWrite,
 {
-    fut: LocalBoxFuture<'static, Result<SslStream<T>, HandshakeError<T>>>,
-    _guard: CounterGuard,
+    Init(Option<Result<SslStream<T>, Error>>, Option<CounterGuard>),
+    Accept(Option<SslStream<T>>, Option<CounterGuard>),
 }
 
 impl<T: AsyncRead + AsyncWrite + Unpin> Future for AcceptorServiceResponse<T> {
-    type Output = Result<SslStream<T>, HandshakeError<T>>;
+    type Output = Result<SslStream<T>, Error>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let io = futures_util::ready!(Pin::new(&mut self.fut).poll(cx))?;
-        Poll::Ready(Ok(io))
+        loop {
+            match self.as_mut().get_mut() {
+                AcceptorServiceResponse::Init(res, guard) => {
+                    let guard = guard.take();
+                    let stream = res.take().unwrap()?;
+                    let state = AcceptorServiceResponse::Accept(Some(stream), guard);
+                    self.as_mut().set(state);
+                }
+                AcceptorServiceResponse::Accept(stream, guard) => {
+                    ready!(Pin::new(stream.as_mut().unwrap()).poll_accept(cx))?;
+                    // drop counter guard a little early as the accept has finished
+                    guard.take();
+
+                    let stream = stream.take().unwrap();
+                    return Poll::Ready(Ok(stream));
+                }
+            }
+        }
     }
 }